// generated/models/ApiCollectRuleTarget.ts
// @ts-nocheck
/**
 * 6529.io API
 * JSON REST API for 6529.io. New here? A terminology overview and a step-by-step authentication walkthrough (guides &amp; auth) live at <a href=\"https://6529.io/tools/api\">https://6529.io/tools/api</a>. The raw machine-readable spec is downloadable at <a href=\"/openapi.yaml\">/openapi.yaml</a> and <a href=\"/openapi.json\">/openapi.json</a>.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class ApiCollectRuleTarget {
    'asset_key': string;
    'target_quantity': string;
    'maximum_unit_price_wei': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "asset_key",
            "baseName": "asset_key",
            "type": "string",
            "format": ""
        },
        {
            "name": "target_quantity",
            "baseName": "target_quantity",
            "type": "string",
            "format": ""
        },
        {
            "name": "maximum_unit_price_wei",
            "baseName": "maximum_unit_price_wei",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ApiCollectRuleTarget.attributeTypeMap;
    }

    public constructor() {
    }
}